+++ CHANGELOG.md
@@ -15,7 +15,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Fix: Added Built in CT Html Page layout #2321 [SchauDK]
-- Fix: Handle CreateGroupEx with SiteStatus = 1 (provisioning status)
+- Fix: Handle GroupSiteManager/CreateGroupEx with SiteStatus = 1 (provisioning status)
+- Fix: Handle SPSiteManager/Create with SiteStatus = 1 (provisioning status)
 - Fix: Increase default timeout for ALM API calls
 
 ## [3.11.1907.0 - July 2019 release]